feat: add impl for eth_getTransasctionByBlockAndIndex
Transaction Index and specify index will be out of sync as it make no
sense for shardeum. Shardeum does not reorder tx. Tx reach immediate
finality in shardeum - first come first processed. This mean transactions are naturally ordered
by timestamp ascendingly within the block. first index [0] mean the tx
is injected and processed earliest and [arr.lenght - 1] mean tx is injected and processed last
within particular block. With that logic we get the block specify the
RPC consumer and then we plug the index specified into the array and
returns.